backend/requirements.txt
beautifulsoup4==4.13.4
flask==2.0.3
flask-cors==3.0.10
httpx==0.28.1
pypdf2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2025.2
requests==2.26.0
sentence-transformers==2.2.2
supabase==2.15.0
unidecode==1.2.0
werkzeug==2.0.3
redis==5.0.1
